requirements.txt
torch
transformers
datasets
tqdm
Pillow
requests
soundfile
numpy